Add property (needs renamed) to get a dictionary of the tasks in the local
* LocalCache.cs: Add property (needs renamed) to get a dictionary of
the tasks in the local database (much like the dictionary received
from the backend). Fix some indentation.
* SyncManager.cs: Add logic to (1) send local tasks to the remote
backend if they don't exist at the remote backend and (2) update the
local task that was sent to the remote backend with an external ID (so
it won't be sent to the remote backend again.  Currently crashing when
calling the backend's CreateTask method.  Needs work.
* Task.cs: Add ExternalId property (with a getter and setter) that can
get the external ID in the local database or set it.

svn path=/branches/new_cache/; revision=205